--- a/.github/workflows/CI.yml
+++ b/.github/workflows/CI.yml
@@ -5,6 +5,7 @@ on:
     branches:
       - "master"
       - "maintenance/.*"
+      - "lily/*"
   pull_request:
     branches:
       - "master"