--- a/.github/workflows/CI.yml
+++ b/.github/workflows/CI.yml
@@ -5,7 +5,6 @@ on:
     branches:
       - "master"
       - "maintenance/.*"
-      - "lily/*"
   pull_request:
     branches:
       - "master"