--- a/openff/toolkit/utils/rdkit_wrapper.py
+++ b/openff/toolkit/utils/rdkit_wrapper.py
@@ -1111,13 +1111,15 @@ class RDKitToolkitWrapper(base_wrapper.ToolkitWrapper):
         ).reshape(-1, 1)
 
         # Build an exclusion list for 1-2 and 1-3 interactions.
-        excluded_x, excluded_y = zip(*{
-            *[(bond.atom1_index, bond.atom2_index) for bond in molecule.bonds],
-            *[
-                (angle[0].molecule_atom_index, angle[-1].molecule_atom_index)
-                for angle in molecule.angles
-            ],
-        })
+        excluded_x, excluded_y = zip(
+            *{
+                *[(bond.atom1_index, bond.atom2_index) for bond in molecule.bonds],
+                *[
+                    (angle[0].molecule_atom_index, angle[-1].molecule_atom_index)
+                    for angle in molecule.angles
+                ],
+            }
+        )
 
         # Build the distance matrix between all pairs of atoms.
         coordinates = conformer.value_in_unit(unit.angstrom)
@@ -1127,10 +1129,10 @@ class RDKitToolkitWrapper(base_wrapper.ToolkitWrapper):
         # np.einsum is both faster than np.diag, and not read-only
         diag = np.einsum("ii->i", ab)
         # modifying in-place lets us take advantage of diag
-        ab += (ab - diag - diag[..., np.newaxis])
+        ab += ab - diag - diag[..., np.newaxis]
         # Handle edge cases where the squared distance is slightly negative due to
         # precision issues
-        diag[:] = -0.
+        diag[:] = -0.0
         distances = np.sqrt(-ab)
 
         inverse_distances = np.reciprocal(