@lilyminium 2021-09-13
> I'll take you on your word that the RuntimeWarning because of precision errors don't show up any more. I don't get it on openff/toolkit/tests/test_toolkits.py::TestRDKitToolkitWrapper::test_elf_compute_electrostatic_energy without this patch. It certainly seems like they'd be gone because of the order of operations here, and you're the person who wanted them gone.

I don't get it for the test case either, but for real molecules I do. For the long hydrocarbon string I used in the big molecule example:

<img width="991" alt="Screen Shot 2021-09-13 at 3 30 18 pm" src="https://user-images.githubusercontent.com/31115101/133164801-2fae7888-55d1-4328-93ef-110dd9eccb3d.png">


I saved the values to `test.txt`:

```conformer = np.load("test.txt")```

[test.txt](https://github.com/openforcefield/openff-toolkit/files/7157676/test.txt)