@mattwthompson 2021-09-13
Approval stands without double-checking every detail but it's always nice to see patches like this squash erroneous warnings:

```
In [10]: conformer = np.loadtxt('/Users/mwt/Downloads/test.txt') * unit.angstrom

In [11]: old_coords(conformer)
<ipython-input-4-6bd55376c816>:2: RuntimeWarning: invalid value encountered in sqrt
  distances = np.sqrt(
Out[11]:
array([[ 0.        ,  1.52017794,  2.5206353 , ..., 35.3855318 ,
        34.57575288, 34.56935347],
       [ 1.52017794,  0.        ,  1.52947889, ..., 34.14586922,
        33.30422763, 33.29825586],
       [ 2.5206353 ,  1.52947889,  0.        , ..., 32.86492709,
        32.05743949, 32.05102204],
       ...,
       [35.3855318 , 34.14586922, 32.86492709, ...,  0.        ,
         1.77518397,  1.7718458 ],
       [34.57575288, 33.30422763, 32.05743949, ...,  1.77518397,
         0.        ,  1.77168836],
       [34.56935347, 33.29825586, 32.05102204, ...,  1.7718458 ,
         1.77168836,  0.        ]])

In [12]: new_coords(conformer)
Out[12]:
array([[ 0.        ,  1.52017794,  2.5206353 , ..., 35.3855318 ,
        34.57575288, 34.56935347],
       [ 1.52017794,  0.        ,  1.52947889, ..., 34.14586922,
        33.30422763, 33.29825586],
       [ 2.5206353 ,  1.52947889,  0.        , ..., 32.86492709,
        32.05743949, 32.05102204],
       ...,
       [35.3855318 , 34.14586922, 32.86492709, ...,  0.        ,
         1.77518397,  1.7718458 ],
       [34.57575288, 33.30422763, 32.05743949, ...,  1.77518397,
         0.        ,  1.77168836],
       [34.56935347, 33.29825586, 32.05102204, ...,  1.7718458 ,
         1.77168836,  0.        ]])